refactor(app): move `TopLevelScreen` to `ui` package
